♻️ Refactor PostPreview component
