--- src/components/PostPreview/PostPreview.scss
@@ -1,26 +1,0 @@
-.PostPreview {
-  padding: 3rem 0;
-
-  & + & {
-    border-top: 1px solid var(--separator-color);
-  }
-
-  &:nth-last-of-type() {
-    border-bottom: 1px solid var(--separator-color);
-  }
-}
-
-.PostPreview__title {
-  a {
-    color: var(--title-color);
-  }
-}
-
-.PostPreview__meta {
-  color: var(--muted-text-color);
-  font-size: var(--font-size-small);
-}
-
-.PostPreview footer {
-  margin-left: -0.6rem;
-}
